+++ src/main/java/me/marin/lockout/lockout/DefaultGoalRegister.java
@@ -24,6 +24,8 @@ import me.marin.lockout.lockout.goals.have_more.HaveMostAdvancementsGoal;
 import me.marin.lockout.lockout.goals.have_more.HaveMostUniqueCraftsGoal;
 import me.marin.lockout.lockout.goals.have_more.HaveMostXPLevelsGoal;
 import me.marin.lockout.lockout.goals.have_more.HaveMostHoppersGoal;
+import me.marin.lockout.lockout.goals.have_more.HaveMostLeaflitterGoal;
+import me.marin.lockout.lockout.goals.have_more.HaveMostDiamondBlocksGoal;
 import me.marin.lockout.lockout.goals.kill.*;
 import me.marin.lockout.lockout.goals.kill.unique.Kill10UniqueHostileMobsGoal;
 import me.marin.lockout.lockout.goals.kill.unique.Kill13UniqueHostileMobsGoal;
@@ -384,6 +386,8 @@ public class DefaultGoalRegister {
         INSTANCE.register(GoalType.HAVE_MOST_PLAYER_KILLS, HaveMostPlayerKillsGoal.class, GoalRequirements.TEAMS_GOAL);
         INSTANCE.register(GoalType.HAVE_MOST_ADVANCEMENTS, HaveMostAdvancementsGoal.class, GoalRequirements.TEAMS_GOAL);
         INSTANCE.register(GoalType.HAVE_MOST_HOPPERS, HaveMostHoppersGoal.class, GoalRequirements.TEAMS_GOAL);
+        INSTANCE.register(GoalType.HAVE_MOST_LEAFLITTER, HaveMostLeaflitterGoal.class, GoalRequirements.TEAMS_GOAL);
+        INSTANCE.register(GoalType.HAVE_MOST_DIAMOND_BLOCKS, HaveMostDiamondBlocksGoal.class, GoalRequirements.TEAMS_GOAL);
         INSTANCE.register(GoalType.HAVE_YOUR_SHIELD_DISABLED, HaveShieldDisabledGoal.class);
         INSTANCE.register(GoalType.ITEM_FRAME_IN_ITEM_FRAME, ItemFrameInItemFrameGoal.class);
         INSTANCE.register(GoalType.FILL_CAMPFIRE, FillCampfireWithFoodGoal.class);