+++ src/main/java/me/marin/lockout/lockout/GoalType.java
@@ -246,5 +246,7 @@ public class GoalType {
     public static final String FILL_SHELF = "FILL_SHELF";
     public static final String KILL_BLAZE_WITH_SNOWBALL = "KILL_BLAZE_WITH_SNOWBALL";
     public static final String HAVE_MOST_HOPPERS = "HAVE_MOST_HOPPERS";
+    public static final String HAVE_MOST_LEAFLITTER = "HAVE_MOST_LEAFLITTER";
+    public static final String HAVE_MOST_DIAMOND_BLOCKS = "HAVE_MOST_DIAMOND_BLOCKS";
 
 }