+++ src/main/java/me/marin/lockout/server/handlers/EndServerTickEventHandler.java
@@ -5,6 +5,8 @@ import me.marin.lockout.LockoutRunnable;
 import me.marin.lockout.lockout.Goal;
 import me.marin.lockout.lockout.goals.have_more.HaveMostXPLevelsGoal;
 import me.marin.lockout.lockout.goals.have_more.HaveMostHoppersGoal;
+import me.marin.lockout.lockout.goals.have_more.HaveMostLeaflitterGoal;
+import me.marin.lockout.lockout.goals.have_more.HaveMostDiamondBlocksGoal;
 import me.marin.lockout.lockout.goals.misc.EmptyHungerBarGoal;
 import me.marin.lockout.lockout.goals.misc.ReachBedrockGoal;
 import me.marin.lockout.lockout.goals.misc.ReachHeightLimitGoal;
@@ -63,6 +65,26 @@ public class EndServerTickEventHandler implements ServerTickEvents.EndTick {
                     lockout.playerHopperCounts.put(player.getUuid(), hopperCount);
                 }
                 lockout.recalculateHoppersGoal(goal);
+            }
+
+            if (goal instanceof HaveMostLeaflitterGoal) {
+                for (ServerPlayerEntity player : server.getPlayerManager().getPlayerList()) {
+                    if (!lockout.isLockoutPlayer(player.getUuid())) continue;
+                    
+                    int leaflitterCount = player.getInventory().count(Items.LEAF_LITTER);
+                    lockout.playerLeaflitterCounts.put(player.getUuid(), leaflitterCount);
+                }
+                lockout.recalculateLeaflitterGoal(goal);
+            }
+
+            if (goal instanceof HaveMostDiamondBlocksGoal) {
+                for (ServerPlayerEntity player : server.getPlayerManager().getPlayerList()) {
+                    if (!lockout.isLockoutPlayer(player.getUuid())) continue;
+                    
+                    int diamondBlockCount = player.getInventory().count(Items.DIAMOND_BLOCK);
+                    lockout.playerDiamondBlockCounts.put(player.getUuid(), diamondBlockCount);
+                }
+                lockout.recalculateDiamondBlocksGoal(goal);
             }
 
             if (goal.isCompleted()) continue;